Fix test fixture for duplicate-names test
Although this test asserts an error that happens pretty early, if we were to revert
the behavior that caused this error, this test fixture monorepo would reveal itself to
be incorrect. Our code currently requires a lockfile. In order to create a lockfile,
with pnpm, we need to have a dependency. So this commit adds an in-repo dependency
for the purpose of generating a lockfile, so that this is a 'valid' monorepo to our
code