Refactor Task Definition retrieval to support composable configs(#3405)
In order to create composable configs, we need to calculate Task
dependencies and topological dependencies when iterating per-workspace.
Currently this happens too early. 

This commit moves part of the work happening when building the TaskGraph
to a later stagey where we are prepping to create PackageTasks.

This PR moves TopoDeps and Persistent values, because there is no
business logic around them. In a future PR, we will also move `Deps`.