Refactor task dependencies retrieval to support composable configs (#3410)
This means we look up a task's dependencies when iterating
per-workspace, rather than per-task. This will allow us to customize
each task per workspace based on additional config that may be present
in that workspace. After this change, `core.Task` is essentially a tuple that
contains the Task name and the Task Definition. 

This follows the work in 2a8832b4b that moved other parts of TaskDefinition
retrieval to per-workspace.