Store TaskDefinitions in the Graph by Task ID (#3428)
Moves TaskDefinitions in the `CompleteGraph` struct, so we can look it
up during execution. We want to do this here, because _during_ execution,
The TaskGraph has already been constructed, so it is too late to change the definition
at the point. Although this commit does not try to change any TaskDefinitions, we will
implement that for composable configs in the future.