Introduce ResolvedTaskDefinition struct and revert pointer in TaskDefinition
A string-rendered version of TaskDefinition is used to calculate
global hashes. If we use a pointer in the struct, the memory address
is part of the string, making the hash calculation unstable. This
reverts TaskDefinition to what it used to be, and uses ResolvedTaskDefinition
instead to merge configs in.

Also introduces a Meta field in TaskDefinition to keep track of whether
we saw the field from the underlying config or if it's using default assignments.