fix: Skip over repo root package for package inferrence (#3503)
Once again map iteration is the culprit. The failure of `infer_pkg.t` is
caused by package inference iterating over the root package before the
correct candidate package. This results in `pkgPath == repoRoot` and
since `fullInferencePath` is necessarily contained in the repo root, the
root package will always be selected as the exact package. The fix is
just skipping over the root package.

First commit adds a root package to the tests. The package inference
cases will fail in a non-deterministic fashion since this bug depends on
the repo root package being iterated over before the package that should
be inferred. The next commit is a fix.