[FIX]: VSCode "Parsing error: Cannot find module 'next/babel" PNPM ESLint (#3489)
# 🪲 Error Reported

```bash
Parsing error: Cannot find module 'next/babel'
```

# 💻 Where does the Error Spawn?

Requirements:

1. PNPM
2. ESLint
3. Monorepo (Turborepo)
4. Next.js

If all of the above requirements ✅ checked, you can see the Error in
every ESLint config file in the children directories.

# 🤔 Cause

Still needed to figure out why does this happen. My hypothesis as it is
because of the PNPM node_modules design that is read by VSCode's ESLint
extension. But it's because ESLint cannot resolve `next/babel`.

# 🚀 How this solution works?

This works because `require.resolve` uses the node's module resolution
to return the absolute path to a module. It can provide that absolute
path to any importing modules.