fix: pnpm prune misses root dependency (#3483)
Fixes #2798 which comprised of two issues:
- We eagerly checked if a version specifier was already resolved in the
package list. This can lead to grabbing a version of the package that
doesn't have the expected peer dependencies. This was as easy as moving
this check after we have gone down the happy path.
- Not copying internal dependencies of the monorepo root. We now
internally add `//` to the prune scope to make sure we get any internal
dependencies.

For reviewers: I highly recommend reviewing the first commit in this PR
initially as it unifies the pnpm lockfile implementations. I realized
that my code split was unnecessary and made fixing bugs for lockfile v5
and v6 very difficult.