Convert tests for persistent task validation to prysk
These tests make heavy use of the internals of setting up a TaskGraph.
We want to rework how that happens to support lazily loading task
definitions. Instead of rewriting all these unit tests, this commit
updates the tests to integration tests, so we don't have to worry about
the internals. This will also make the tests resilient to the Rust port.