fix(use correct serialization logic for turbo.json) (#3580)
Fixes #3570 

The issue here was that we implemented the custom marshalling behavior
on a *pointer* of `TaskDefinition`, but the pipeline is
`map[string]TaskDefinition`. I chose to change the `MarshalJSON`
implementation to not be on a pointer type. I also added custom
marshalling behavior to output mode so we no longer output the number of
the enum.