Capture engine/run unit tests with prysk (#3592)
This converts unit tests exercising the task graph to integration tests. 
The purpose of this change is to make it easier to land lazy loading task definitions,
which breaks all these unit tests, due to their use of the AddTask API to
setup the test. Additionally, the lazy loading feature will require that these
tests have access to the file system to read from config file, which would
make them closer to integration tests already.

Note that the new tests are here are:

- not 1:1 with the old ones, but attempt to capture the scenarios in spirit.
- asserted with `--graph`, because parallel execution makes it a bit more challenging to
assert against logs.
- combined with existing topological-deps tests