Add force_post_to_slack to repository dispatch (#3649)
This adds the `force_post_to_slack` option to the Next.js release
workflow's `repository_dispatch` event. It defaults to `false` for the
repository dispatch event and is now required for manual dispatch. This
should ensure that the value is always present as it's expected to be a
boolean later in the workflow.