fix: remove turborepo binary dependency on ffi (#3683)
Currently our release process is broken due to `protoc` not being setup
on the machine:
https://github.com/vercel/turbo/actions/runs/4118079447/jobs/7110161689#step:6:363

We shouldn't need protoc when building the top level Rust binary as it's
only used when building `turborepo-ffi` which is used by `go-turbo` to
call Rust code. `turborepo-ffi` is already built if one runs `make
go-turbo` which happens as part of our Rust binary's build script (when
not creating a release build). During a release build we don't build
`go-turbo` since it is built in an earlier step and is only needed when
we package up the two binaries.

We could instead setup `protoc` in the Rust build process, but that adds
time to the release process.