Include workspace turbo.json in inputs (#3778)
When `inputs` is configured for a task, we want to ensure that changes
to the `turbo.json` in that workspace cause a cache miss. This is
similar to how we treat the `package.json` in the workspace, and also
how we treat the root `turbo.json` for the global hash.