Ensure tasks can depend on Workspace-Specific tasks (#3820)
This specifically adds a test to ensure that tasks from Workspace Configs
can still depend on tasks from other workspaces with the worksapce#task
syntax. Because workspace#task syntax is not allowed for pipeline entries,
we need to make sure we don't accidentally start parsing all task references
and validating them this way.