feat(eslint): support turbo.json workspace configs (#3846)
Support turbo.json workspace configs. 

Still to do:
- [x] Support workspace specific linting

There exists one edge case here, if a user forgets to add an "extend" to workspace config, we _can_ treat it as root. I punted on this one for now since it's not this packages domain to solve that (at the moment)