Add --log-prefix=none option (#3851)
This adds a `--log-prefix=none` CLI argument to `turbo run` commands
that will remove the `<package>:<task>: ` prefix when running tasks
(both on execute and replay from cache). This removal will also apply to
"single package" repos where the prefix is just `<task>: `

This new option is intended to alleviate some user issues piping logs
into other tools, but comes with a Buyer Beware warning that logs from
parallel tasks are still interleaved. We may implement a buffering
mechanism (long time issue: #219), that will further help with this, but
this PR does not do that.