Add ExpandedInputs to Dry Run Summary
Give the CompleteGraph struct access to the TaskHashTracker instance
so we can use it while constructing PackageTasks. PackageTasks are
then available when walking the graph during dry run execution, and
we can add them into the Dry Run output. This approach also makes it
trivial to use the PackageTask.ExpandedInputs during RealRun execution
so we can print them as debug output in a Run Summary