feat(turbo-utils): bundle to esm (#3952
Now that we're throwing a few more things in turbo-utils, we should
bundle it to esm instead of cjs to support code splitting. This is an
internal package, and everything that consumes this bundles to cjs, so
compatibility isn't an issue, and this will ensure each package that
uses a util from turbo-utils only gets what it needs instead of the
entire package.

I tested this locally to see the bundle size difference. 

For turbo-ignore, the bundle size drops from `245K` to `229K`. No _huge_
(~6%) but worth it.