Remove non-functional css property
A typographical error in a css property resulted in extraneous css.

This commit removes the property in question for housekeeping the style sheet.

Even if correct, I believe the property would have had no effect on rendering or side effect with run-time style modification.  The property has a value of `0` and is not marked important.

Alternatives at discretion of maintainers:
  * Fix the typo `transfom` -> `transform`. 

This solution was tested locally and had no effect on document render at each of the project's breakpoints in Chrome, Firefox, and Safari (latest releases 26FEB2023), But was not tested on other browsers or mobile.  I selected the removal option as a result of this incomplete testing.

Versioning / Stability / Urgency:
Does not address critical, breaking, or security issues.  Low priority.