Root inference: skip turbo.json (#3946)
Integration tests for inferring the root. This demonstrates that detection based upon `turbo.json` (and the change made in #3938) is not (particularly) relevant to root identification.

How to review:
- Each commit is pretty self-contained. You can see me adding a debug line to assert against in the first commit.
- The third commit is both a revert of #3938 and a full deletion of the `turbo.json`-included detection logic.

<details>
<summary>OUTDATED</summary>
There is _one_ behavior change. If you previously:
1. Had nested workspaces (which is not supported by us or any package manager).
2. Had a `turbo.json` at the "outer" one.
3. Did _not_ have a `turbo.json` at the "inner" one

Then we would have selected the outer one. We now select the inner one, and throw an error because the inner one does not have a `turbo.json`. I believe this to be an _improvement_.
</details>

***

#3938 treats any `turbo.json` that has comments in it as nonexistent since `serde_json` doesn't support JSONC. That everything still worked with parsing that is throwing an error made it clear that the `turbo.json` check was not being used except for the described edge case.