feat(codemod): support workspace configs (#3992)
This fixes turbo codemod not building due to the new config schema introduced for workspace configs. 

It also adds workspace config support and tests to the codemods that are relevant (env vars and outputs). This isn't strictly necessary as workspace configs were introduced after the version that required these codemods, but I wanted to add this just in case someone uses an old example from around the web to update a workspace config or similar.