Introduce FS::track to track without reading a file
Whenever we write a file, we also read its contents to see if we need to write at all (maybe we can skip the sys call). We want to ensure the file always contains the updated contents (imagine a user edits the file). But by reading the file into a RopeVc, we're doubling the memory usage of every write (one copy for the content of the write, one copy for the read that we then compare against).

By tracking without reading, we can maybe free the read's RopeVc in the future.