Add chotot to showcase (#4097)
We use Turporepo at Chợ Tốt to manage internal libraries and build one
of the highest traffic websites on Chợ Tốt. The amount of time that
Turporepo has helped us save is enormous, and we are very grateful to
you for creating such a powerful tool and sharing it with the world
completely for free.