Add static_embed feature
Based on https://vercel.slack.com/archives/C04KC8A53T7/p1678496137453779?thread_ts=1678468241.013549&cid=C04KC8A53T7, our Next.js test suite is failing because the default test compilation is not portable. This is because the `embed_dir` and `embed_file` statically link the directory path into the binary, but do not statically embed the contents. If this same binary is executed on another machine that doesn't follow the exact same directory structure, it'll fail to embed the internal assets.