Revert #3542 (#4197)
### Description

The fix introduced in #3542 was a workaround for CSS in dynamic imports
being included multiple times in a page, potentially overriding
precedence of earlier chunks, but introduced its own precedence issue
(the semantics of `@layer` are not fully compatible with our use case).
This is no longer necessary since #4056 made it so dynamic imports no
longer include items from the parent chunk.

### Testing Instructions

create-next-app with app support now looks visually identical with and
without --turbo.

Snapshots.