Switch to testing::fixture
The old `test_generator::test_resources` could only find resources relative to the project root, and not relative to the test file. Because of the new nextpack directory structure, this causes all kinds of headaches between running in nextpack and running in turbo (`crates/turbopack-tests` vs `nextpack/turbo-crates/crates/turbopack/tests`).

We'll need to do the same in `next-dev-tests` in Next.js repo.