Bring back chunk runner logic for the none runtime (#4285)
This is a fix on top of #4141 to ensure that all chunks instantiate in
the proper order when running on the Next.js edge runtime. This brings
back a similar "runner" logic as what we had before #4141, where
registering a new chunk might cause other chunks to be instantiated if
we consider all their dependency chunks to be registered.