Bump rustc version. (#4331)
The [build
issues](https://github.com/vercel/turbo/actions/workflows/turborepo-release-step-2.yml)
appear to be related to the `rustc` version. Rust is in the process of
attempting to land `llvm@16` which was released on March 18 (depending
on where you stand in the world).

I've bisected the release versions and confirmed that
`nightly-2023-03-21` is the last working release— _coincidentally_ the
version from the day we ran the 1.8.5 release. For now I'm hardcoding
that into our release process; I have not done any further investigation
of the actual root cause which will enable us to update past
`nightly-2023-03-21`.

We will need to do that at some point in order to make sure we don't
lose the ability to release.