ci(workflow): resurrect daily next.js test (#4287)
### Description

Part 1 for WEB-762.

This is an attempt to simple-reenablement to previous next.js
integration test.

Please note the general test strategy will be still revisited: i.e there
was a discussion for making next.js upstream test to turbopack-aware,
and some other approaches as well. Meanwhile, this'll allow to resurrect
old dashboard / slack status as quick as possible. Once we have stable
ground for the whole test, this can go away then.