feat(place patches in json folder when pruning for docker) (#4318)
### Description

Implements #4289

Using the `--docker` flag with prune produces a `full` and `json`
directory where the `json` directory only includes everything required
to install dependencies. If a dependency is patched it makes sense to
include the patch in the `json` directory as both the `package.json` and
the lockfile reference the patch file.

### Testing Instructions

Added an integration test to verify that patches end up in the `json`
output folder.